pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "python-dotenv>=1.1.1",
]
